Create a class to manage list
